fix: v1.4.2编译错误修复 - nullable String安全调用, LocalBroadcastManager导入, offset类型转换

+++ app/src/main/java/com/radio/app/services/SubtitleGeneratorService.kt
@@ -288,7 +288,7 @@ class SubtitleGeneratorService : Service() {
                                     this.confidence = 0.5
                                 }
                                 callback.onSubtitleGenerated(t)
-                                lastPartialOutput = offset
+                                lastPartialOutput = offset.toLong()
                                 Log.d(TAG, "Vosk partial: $startSec-$endSec: $partialText")
                             }
                         } catch (_: Exception) {
